requirements.txt
python-dateutil
fpdf2==2.8.2
